+++ changelog.md
@@ -1,3 +1,8 @@
+## v0.6 ##
+* Fixed bug in Numba version where parallel construction of pre-computing splits failed.
+* Fixed bug in Numba version where variance of distributions (other than Normal) was not properly clipped.
+* Fixed Gamma distribution in Numba version
+
 ### v0.5.1 ###
 * Fixed bug in PyPi release where the custom CUDA kernel was not included in the distribution.